backend/requirements.txt
fastapi
uvicorn
python-socketio
websockets
qwen-agent
asyncio
typing
pydantic
python-multipart